+++ src/app/layout.tsx
@@ -19,6 +19,7 @@ import CookieConsentProvider from "@/browser-lib/cookie-consent/CookieConsentPro
 import { FAVICON_LINKS_HEAD_INNER_HTML } from "@/image-data";
 import getBrowserConfig from "@/browser-lib/getBrowserConfig";
 import { OakNotificationsProvider } from "@/context/OakNotifications/OakNotificationsProvider";
+import { SaveCountProvider } from "@/context/SaveCount/SaveCountProvider";
 
 // https://nextjs.org/docs/app/getting-started/metadata-and-og-images
 export const metadata: Metadata = {
@@ -88,9 +89,9 @@ export default function RootLayout({
                     }}
                   >
                     <AnalyticsWrapper>
-                      <OakBox $width="100vw" $height="100vh">
-                        {children}
-                      </OakBox>
+                      <SaveCountProvider>
+                        <OakBox $height="100vh">{children}</OakBox>
+                      </SaveCountProvider>
                     </AnalyticsWrapper>
                   </ClerkProvider>
                 </OakNotificationsProvider>