+++ src/components/AppComponents/TopNav/TeachersTopNavHamburger/HamburgerMainMenu.tsx
@@ -87,7 +87,7 @@ function SubjectsSection(
           </OakHeading>
           <OakSvg
             $position={"absolute"}
-            $color={"mint"}
+            $color={"bg-decorative1-main"}
             $display={"block"}
             name={"underline"}
             $height={"spacing-8"}